pyproject.toml
[project]
name = "ragsworth"
version = "0.1.0"
description = "An embeddable AI chatbot with RAG capabilities and multi-provider LLM support"
authors = []
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "spacy>=3.8.0",
    "openai>=1.12.0",
    "anthropic>=0.19.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "python-multipart>=0.0.9",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    "aiohttp>=3.9.0",
    "tqdm>=4.66.0",
    "pymilvus>=2.3.0,<2.5.0",
    "environs>=9.5.0,<14.0.0",
    "marshmallow>=3.19.0,<4.0.0",
    "backports.tarfile>=0.1.1",
    "click>=8.1.8",
    "mock>=5.2.0",
    "pypdf>=5.5.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "passlib>=1.7.4",
    "argon2-cffi>=23.1.0",
    "jwt>=1.3.1",
    "pyjwt>=2.10.1",
]

[project.scripts]
ragsworth = "ragsworth.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=ragsworth --cov-report=term-missing" 

[tool.hatch.build.targets.wheel]
packages = ["ragsworth"] 
